+++ sdk/storage/azure-storage-blob/azure/storage/blob/__init__.py
@@ -26,6 +26,8 @@ from ._shared.models import (
     UserDelegationKey,
     Services,
 )
+from ._shared.session import SessionProvider
+from ._shared.session_provider import ContainerSessionProvider
 from ._generated.models import RehydratePriority
 from ._models import (
     BlobType,
@@ -62,6 +64,7 @@ from ._models import (
     ObjectReplicationPolicy,
     ObjectReplicationRule,
     ImmutabilityPolicy,
+    SessionOptions,
 )
 from ._list_blobs_helper import BlobPrefix
 
@@ -265,4 +268,7 @@ __all__ = [
     "ObjectReplicationPolicy",
     "ObjectReplicationRule",
     "Services",
+    "SessionProvider",
+    "SessionOptions",
+    "ContainerSessionProvider",
 ]