+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/_container_client_async.py
@@ -119,6 +119,11 @@ class ContainerClient(  # type: ignore [misc]  # pylint: disable=too-many-public
     :keyword str audience: The audience to use when requesting tokens for Azure Active Directory
         authentication. Only has an effect when credential is of type TokenCredential. The value could be
         https://storage.azure.com/ (default) or https://<account>.blob.core.windows.net.
+    :keyword bool use_session: If True, enable session-based authentication for this container.
+        When enabled, eligible GET requests issued by this client will be authenticated using
+        a short-lived session credential obtained from the service instead
+        of the provided AsyncTokenCredential. Only supported with an AsyncTokenCredential;
+        ValueError is raised otherwise. Defaults to False.
 
     .. admonition:: Example: